% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectory_analysis.R
\name{calculate_sector_trajectories_vectorized}
\alias{calculate_sector_trajectories_vectorized}
\title{Vectorized sector trajectory calculation helper function}
\usage{
calculate_sector_trajectories_vectorized(
  dt,
  quarters_dt,
  start_col,
  end_col,
  person_col,
  arco_col,
  sector_col,
  reference_sectors,
  quarter_days
)
}
\arguments{
\item{dt}{Employment data}

\item{quarters_dt}{Quarter data for all persons}

\item{start_col}{Start date column name}

\item{end_col}{End date column name}

\item{person_col}{Person ID column name}

\item{arco_col}{Arco column name for employment status}

\item{sector_col}{Sector column name}

\item{reference_sectors}{data.table with reference sectors per person}

\item{quarter_days}{Number of days per quarter}
}
\value{
data.table with sector trajectories
}
\description{
Vectorized sector trajectory calculation helper function
}
\keyword{internal}
