% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/career_clustering.R
\name{.compute_calinski_harabasz}
\alias{.compute_calinski_harabasz}
\title{Compute Calinski-Harabasz Index (Internal)}
\usage{
.compute_calinski_harabasz(features_matrix, clusters)
}
\arguments{
\item{features_matrix}{Numeric matrix}

\item{clusters}{Vector of cluster assignments}
}
\value{
Numeric CH index value
}
\description{
Compute Calinski-Harabasz Index (Internal)
}
\keyword{internal}
