% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/career_clustering.R
\name{.compute_distance_to_centroid}
\alias{.compute_distance_to_centroid}
\title{Compute Distance to Cluster Centroid (Internal)}
\usage{
.compute_distance_to_centroid(
  features_matrix,
  cluster_ids,
  cluster_result,
  batch_size = 1e+05
)
}
\arguments{
\item{features_matrix}{Numeric feature matrix}

\item{cluster_ids}{Vector of cluster assignments}

\item{cluster_result}{Clustering result object with centers}

\item{batch_size}{Integer. Batch size for processing (reduces memory spikes)}
}
\value{
Numeric vector of distances
}
\description{
Compute Distance to Cluster Centroid (Internal)
}
\keyword{internal}
