% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/career_clustering.R
\name{.compute_dunn_index}
\alias{.compute_dunn_index}
\title{Compute Dunn Index (Internal)}
\usage{
.compute_dunn_index(features_matrix, clusters)
}
\arguments{
\item{features_matrix}{Numeric matrix}

\item{clusters}{Vector of cluster assignments}
}
\value{
Numeric Dunn index value
}
\description{
Compute Dunn Index (Internal)
}
\keyword{internal}
