% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/career_clustering.R
\name{.determine_optimal_clusters}
\alias{.determine_optimal_clusters}
\title{Determine Optimal Number of Clusters (Internal)}
\usage{
.determine_optimal_clusters(
  features_matrix,
  method,
  min_k = 3,
  max_k = 6,
  min_cluster_size = 10,
  nstart = 25,
  seed = 123,
  verbose = FALSE,
  use_sampling = FALSE,
  sample_size = 50000
)
}
\arguments{
\item{features_matrix}{Numeric matrix of features}

\item{method}{Clustering method}

\item{min_k}{Minimum number of clusters}

\item{max_k}{Maximum number of clusters}

\item{min_cluster_size}{Minimum cluster size}

\item{nstart}{Number of random starts for k-means}

\item{seed}{Random seed}

\item{verbose}{Logical}

\item{use_sampling}{Logical. Use sampling for large datasets?}

\item{sample_size}{Integer. Sample size for silhouette analysis}
}
\value{
Optimal number of clusters
}
\description{
Determine Optimal Number of Clusters (Internal)
}
\keyword{internal}
