% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/career_clustering.R
\name{.generate_cluster_labels}
\alias{.generate_cluster_labels}
\title{Generate Bilingual Cluster Labels Based on Characteristics (Internal)}
\usage{
.generate_cluster_labels(cluster_profiles, verbose = FALSE)
}
\arguments{
\item{cluster_profiles}{data.table with cluster means}

\item{verbose}{Logical}
}
\value{
data.table with cluster_id, cluster_label_en, cluster_label_it
}
\description{
Generate Bilingual Cluster Labels Based on Characteristics (Internal)
}
\keyword{internal}
