% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/career_clustering.R
\name{.perform_clustering}
\alias{.perform_clustering}
\title{Perform Clustering with Specified Method (Internal)}
\usage{
.perform_clustering(features_matrix, n_clusters, method, nstart, seed)
}
\arguments{
\item{features_matrix}{Numeric matrix}

\item{n_clusters}{Number of clusters}

\item{method}{Clustering method}

\item{nstart}{Number of random starts}

\item{seed}{Random seed}
}
\value{
Clustering result object
}
\description{
Perform Clustering with Specified Method (Internal)
}
\keyword{internal}
