% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/career_clustering.R
\name{.prepare_clustering_data}
\alias{.prepare_clustering_data}
\title{Prepare Data for Clustering (Internal)}
\usage{
.prepare_clustering_data(
  career_metrics,
  clustering_features,
  id_column,
  standardize,
  verbose
)
}
\arguments{
\item{career_metrics}{data.table with career metrics}

\item{clustering_features}{Character vector of feature names}

\item{id_column}{Character name of ID column}

\item{standardize}{Logical}

\item{verbose}{Logical}
}
\value{
List with features_matrix, features_scaled, ids
}
\description{
Prepare Data for Clustering (Internal)
}
\keyword{internal}
