% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/career_clustering.R
\name{.reassign_small_clusters}
\alias{.reassign_small_clusters}
\title{Reassign Small Clusters to Nearest Larger Cluster (Internal)}
\usage{
.reassign_small_clusters(
  cluster_assignments,
  features_matrix,
  cluster_result,
  small_clusters,
  min_cluster_size
)
}
\arguments{
\item{cluster_assignments}{data.table with cluster assignments}

\item{features_matrix}{Numeric feature matrix}

\item{cluster_result}{Clustering result object}

\item{small_clusters}{Vector of small cluster IDs}

\item{min_cluster_size}{Minimum size threshold}
}
\value{
Updated cluster_assignments data.table
}
\description{
Reassign Small Clusters to Nearest Larger Cluster (Internal)
}
\keyword{internal}
