% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/career_clustering.R
\name{.select_clustering_features}
\alias{.select_clustering_features}
\title{Select Clustering Features Based on Category (Internal)}
\usage{
.select_clustering_features(career_metrics, features, verbose = FALSE)
}
\arguments{
\item{career_metrics}{data.table with career metrics}

\item{features}{Character vector of feature categories}

\item{verbose}{Logical}
}
\value{
Character vector of selected feature names
}
\description{
Select Clustering Features Based on Category (Internal)
}
\keyword{internal}
