% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/career_clustering.R
\name{.stratified_cluster_sample}
\alias{.stratified_cluster_sample}
\title{Stratified Sampling Preserving Cluster Proportions (Internal)}
\usage{
.stratified_cluster_sample(clusters, sample_size)
}
\arguments{
\item{clusters}{Vector of cluster assignments}

\item{sample_size}{Integer. Target sample size}
}
\value{
Integer vector of sample indices
}
\description{
Stratified Sampling Preserving Cluster Proportions (Internal)
}
\keyword{internal}
