% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consolidate_employment.R
\name{.use_vecshift_temporal_consolidation_optimized}
\alias{.use_vecshift_temporal_consolidation_optimized}
\title{OPTIMIZED: Use vecshift temporal consolidation functions with performance enhancements}
\usage{
.use_vecshift_temporal_consolidation_optimized(
  dt,
  consolidation_type,
  show_progress
)
}
\arguments{
\item{dt}{Working data.table}

\item{consolidation_type}{Consolidation type}

\item{show_progress}{Show progress messages}
}
\value{
Consolidated data.table
}
\description{
OPTIMIZED: Use vecshift temporal consolidation functions with performance enhancements
}
\keyword{internal}
