% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_visualization.R
\name{validate_animation_setup}
\alias{validate_animation_setup}
\title{Validate Animation Setup}
\usage{
validate_animation_setup()
}
\description{
Internal function to validate that animation prerequisites are met
}
\keyword{internal}
