% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consolidation_metrics.R
\name{print.longworkR_transitions_with_metrics}
\alias{print.longworkR_transitions_with_metrics}
\title{Print method for transitions with metrics}
\usage{
\method{print}{longworkR_transitions_with_metrics}(x, ...)
}
\arguments{
\item{x}{Transitions with metrics object}

\item{...}{Additional arguments}
}
\description{
Print method for transitions with metrics
}
